requirements.txt
streamlit==1.22.0
openpyxl==3.0.10
plotly==5.11.0
numpy==1.22.1
keras==2.10.0
sklearn-crfsuite==0.3.6
jinja2-time==0.2.0
async-timeout==4.0.2
tensorflow==2.10.0
scikit-learn==1.2.2
